refacto some unuseful code
