+++ openaev-front/src/admin/components/atomic_testings/atomic_testing/TargetResultsSecurityPlatform.tsx
@@ -9,7 +9,7 @@ import Drawer from '../../../../components/common/Drawer';
 import { useFormatter } from '../../../../components/i18n';
 import type { InjectExpectationResult, InjectExpectationTrace } from '../../../../utils/api-types';
 import { type InjectExpectationsStore } from '../../common/injects/expectations/Expectation';
-import { getSourceLabel } from '../../simulations/simulation/validation/expectations/ExpectationUtils';
+import { getSourceLabel } from '../../common/injects/expectations/ExpectationUtils';
 
 const useStyles = makeStyles()(() => ({ flexContainer: { display: 'flex' } }));
