+++ openaev-front/src/admin/components/atomic_testings/atomic_testing/target_result/InjectExpectationCard.tsx
@@ -15,8 +15,7 @@ import { emptyFilled } from '../../../../../utils/String';
 import { isAssets } from '../../../../../utils/target/TargetUtils';
 import { PermissionsContext } from '../../../common/Context';
 import type { InjectExpectationsStore } from '../../../common/injects/expectations/Expectation';
-import { isManualExpectation, isTechnicalExpectation } from '../../../common/injects/expectations/ExpectationUtils';
-import { isAgentExpectation, isAssetExpectation, isAssetGroupExpectation, isPlayerExpectation, useIsManuallyUpdatable } from '../../../simulations/simulation/validation/expectations/ExpectationUtils';
+import { isAgentExpectation, isAssetExpectation, isAssetGroupExpectation, isManualExpectation, isPlayerExpectation, isTechnicalExpectation, useIsManuallyUpdatable } from '../../../common/injects/expectations/ExpectationUtils';
 import InjectExpectationContext from '../context/InjectExpectationContext';
 import ExpirationChip from '../ExpirationChip';
 import InjectExpectationAggregatedAgentsView from './InjectExpectationAggregatedAgentsView';
